# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Kundli Calculation Service API
  description: API for calculating and managing Vedic astrological charts
  version: 1.0.0
  contact:
    name: API Support
    email: support@kundliservice.com

servers:
  - url: https://api.kundliservice.com/v1
    description: Production server
  - url: https://staging-api.kundliservice.com/v1
    description: Staging server
  - url: http://localhost:8000/v1
    description: Local development server

security:
  - bearerAuth: []
  - apiKeyAuth: []

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key

  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
          description: Error code
        message:
          type: string
          description: Error message
        details:
          type: object
          description: Additional error details

    BirthDetails:
      type: object
      required:
        - date
        - time
        - latitude
        - longitude
        - timezone
      properties:
        date:
          type: string
          format: date
          example: "2025-01-05"
        time:
          type: string
          format: time
          example: "06:32:16"
        latitude:
          type: number
          format: float
          example: 28.6139
        longitude:
          type: number
          format: float
          example: 77.2090
        timezone:
          type: string
          example: "Asia/Kolkata"

    KundliChart:
      type: object
      properties:
        id:
          type: string
          format: uuid
        birth_details:
          $ref: '#/components/schemas/BirthDetails'
        ascendant:
          type: number
          format: float
        planets:
          type: object
          additionalProperties:
            type: number
            format: float
        houses:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
        aspects:
          type: array
          items:
            type: object
        dashas:
          type: object
        yogas:
          type: array
          items:
            type: object
        created_at:
          type: string
          format: date-time

paths:
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
      responses:
        '201':
          description: User registered successfully
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  token_type:
                    type: string
                    example: "bearer"
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /kundli/calculate:
    post:
      tags:
        - Kundli
      summary: Calculate new Kundli chart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BirthDetails'
      responses:
        '201':
          description: Kundli calculated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KundliChart'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /kundli/{id}:
    get:
      tags:
        - Kundli
      summary: Get Kundli chart by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Kundli chart found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KundliChart'
        '404':
          description: Kundli not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Kundli
      summary: Delete Kundli chart
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Kundli deleted successfully
        '404':
          description: Kundli not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /kundli/validate:
    post:
      tags:
        - Kundli
      summary: Validate birth details
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BirthDetails'
      responses:
        '200':
          description: Birth details are valid
        '400':
          description: Invalid birth details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'